+++ packages/flutter_meedu/example/lib/main.dart
@@ -26,12 +26,12 @@ class MyApp extends StatelessWidget {
           themeMode: controller.themeMode,
           theme: ThemeData.light(),
           darkTheme: ThemeData.dark(),
-          // initialRoute: Routes.SPLASH,
-          home: ProviderWidgetPage(),
+          initialRoute: Routes.SPLASH,
+          // home: ProviderWidgetPage(),
           navigatorObservers: [
             router.observer,
           ],
-          // routes: Pages.routes,
+          routes: Pages.routes,
         ),
       ),
       providers: [