+++ packages/flutter_meedu/example/lib/pages/tabs_page/tabs/home/controller.dart
@@ -1,0 +1,16 @@
+import 'package:meedu/meedu.dart';
+
+class HomeTabController extends SimpleNotifier {
+  final counter = 0.obs;
+
+  @override
+  void onDispose() {
+    counter.close();
+    print("HomeTabController disposed");
+    super.onDispose();
+  }
+}
+
+final homeTabProvider = SimpleProvider(
+  (_) => HomeTabController(),
+);