+++ packages/flutter_meedu/example/lib/routes/pages.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:meedu_example/pages/home/home_page.dart';
 import 'package:meedu_example/pages/login/login_page.dart';
 import 'package:meedu_example/pages/splash/splash_page.dart';
+import 'package:meedu_example/pages/tabs_page/tabs_page.dart';
 
 import 'routes.dart';
 
@@ -11,5 +12,6 @@ abstract class Pages {
     Routes.SPLASH: (_) => SplashPage(),
     Routes.LOGIN: (_) => LoginPage(),
     Routes.HOME: (_) => HomePage(),
+    Routes.TABS: (_) => TabsPage(),
   };
 }