+++ packages/flutter_meedu/example/lib/routes/routes.dart
@@ -2,4 +2,5 @@ abstract class Routes {
   static const SPLASH = '/';
   static const LOGIN = '/login';
   static const HOME = '/home';
+  static const TABS = '/tabs';
 }