+++ packages/meedu/lib/provider.dart
@@ -1,3 +1,4 @@
 library provider;
 
 export 'src/providers/base_provider.dart';
+export 'src/providers/provider_scope.dart';