+++ packages/meedu/lib/src/providers/provider_scope.dart
@@ -1,0 +1,6 @@
+import 'provider_container.dart';
+
+abstract class ProviderScope {
+  /// save the notifiers in one instance of ProviderContainer
+  static final containers = <int, ProviderContainer>{};
+}